feat: create initial migration and repository for users table

+++ src/Lab5/Lab5Project.Application.Models/Users/User.cs
@@ -1,3 +1,3 @@
 namespace Lab5Project.Application.Models.Users;
 
-public record User(long AccountNumber, string Username, long Pin, long Amount);
+public record User(long AccountNumber, string Username, long Pin, decimal Amount);
